requirements.txt
numpy>=1.20
pandas>=1.3
Pillow>=9.0
Django>=4.2,<5.0
